feat: integrate shared package for logging and dependency management across modules
