+++ package.json
@@ -7,7 +7,7 @@
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "typecheck": "tsc --noEmit",
-    "build": "npm -w packages/api run build && npm -w packages/addon run build",
+    "build": "npm -w packages/shared run build && npm -w packages/api run build && npm -w packages/addon run build",
     "start": "npm -w packages/addon start",
     "dev": "npm -w packages/addon run start:dev",
     "dev:cf": "npm run build && npm -w packages/cloudflare-worker run dev",
