+++ packages/addon/package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "easynews-plus-plus-api": "file:../api",
+    "easynews-plus-plus-shared": "file:../shared",
     "parse-torrent-title": "^1.4.0",
     "stremio-addon-sdk": "^1.6.10"
   }
