+++ packages/api/src/api.ts
@@ -1,5 +1,12 @@
 import { EasynewsSearchResponse, FileData, SearchOptions } from './types';
 import { createBasic } from './utils';
+import { createLogger } from 'easynews-plus-plus-shared';
+
+// Create a logger with API prefix and explicitly set the level from environment variable
+export const logger = createLogger({
+  prefix: 'API',
+  level: process.env.EASYNEWS_LOG_LEVEL || undefined, // Use the environment variable if set
+});
 
 export class EasynewsAPI {
   private readonly baseUrl = 'https://members.easynews.com';
@@ -34,24 +41,24 @@ export class EasynewsAPI {
   private getFromCache(cacheKey: string): EasynewsSearchResponse | null {
     const cached = this.cache.get(cacheKey);
     if (!cached) {
-      console.debug(`[EasynewsAPI] Cache miss for key: ${cacheKey.substring(0, 50)}...`);
+      logger.debug(`Cache miss for key: ${cacheKey.substring(0, 50)}...`);
       return null;
     }
 
     const now = Date.now();
     if (now - cached.timestamp > this.cacheTTL) {
-      console.debug(`[EasynewsAPI] Cache expired for key: ${cacheKey.substring(0, 50)}...`);
+      logger.debug(`Cache expired for key: ${cacheKey.substring(0, 50)}...`);
       this.cache.delete(cacheKey);
       return null;
     }
 
-    console.debug(`[EasynewsAPI] Cache hit for key: ${cacheKey.substring(0, 50)}...`);
+    logger.debug(`Cache hit for key: ${cacheKey.substring(0, 50)}...`);
     return cached.data;
   }
 
   private setCache(cacheKey: string, data: EasynewsSearchResponse): void {
-    console.debug(
-      `[EasynewsAPI] Caching ${data.data?.length || 0} results for key: ${cacheKey.substring(0, 50)}...`
+    logger.debug(
+      `Caching ${data.data?.length || 0} results for key: ${cacheKey.substring(0, 50)}...`
     );
     this.cache.set(cacheKey, { data, timestamp: Date.now() });
   }
@@ -71,7 +78,7 @@ export class EasynewsAPI {
       throw new Error('Query parameter is required');
     }
 
-    console.debug(`[EasynewsAPI] Searching for: "${query}" (page ${pageNr}, max ${maxResults})`);
+    logger.debug(`Searching for: "${query}" (page ${pageNr}, max ${maxResults})`);
 
     const cacheKey = this.getCacheKey({
       query,
@@ -114,7 +121,7 @@ export class EasynewsAPI {
     const url = new URL(`${this.baseUrl}/2.0/search/solr-search/advanced`);
     url.search = new URLSearchParams(searchParams).toString();
 
-    console.debug(`[EasynewsAPI] Request URL: ${url.toString().substring(0, 100)}...`);
+    logger.debug(`Request URL: ${url.toString().substring(0, 100)}...`);
 
     try {
       const res = await fetch(url, {
@@ -125,39 +132,37 @@ export class EasynewsAPI {
       });
 
       if (res.status === 401) {
-        console.debug(`[EasynewsAPI] Authentication failed for user: ${this.username}`);
+        logger.debug(`Authentication failed for user: ${this.username}`);
         throw new Error('Authentication failed: Invalid username or password');
       }
 
       if (!res.ok) {
-        console.debug(`[EasynewsAPI] Request failed with status: ${res.status} ${res.statusText}`);
+        logger.debug(`Request failed with status: ${res.status} ${res.statusText}`);
         throw new Error(
           `Failed to fetch search results of query '${query}': ${res.status} ${res.statusText}`
         );
       }
 
       const json = await res.json();
-      console.debug(
-        `[EasynewsAPI] Received ${json.data?.length || 0} results out of ${json.results || 0} total`
-      );
+      logger.debug(`Received ${json.data?.length || 0} results out of ${json.results || 0} total`);
       this.setCache(cacheKey, json);
       return json;
     } catch (error) {
       if (error instanceof Error) {
         if (error.name === 'AbortError') {
-          console.debug(`[EasynewsAPI] Search request timed out for: "${query}"`);
+          logger.debug(`Search request timed out for: "${query}"`);
           throw new Error(`Search request for '${query}' timed out after 20 seconds`);
         }
-        console.debug(`[EasynewsAPI] Error during search: ${error.message}`);
+        logger.debug(`Error during search: ${error.message}`);
         throw error;
       }
-      console.debug(`[EasynewsAPI] Unknown error during search`);
+      logger.debug(`Unknown error during search`);
       throw new Error(`Unknown error during search for '${query}'`);
     }
   }
 
   async searchAll(options: SearchOptions): Promise<EasynewsSearchResponse> {
-    console.debug(`[EasynewsAPI] Starting searchAll for: "${options.query}"`);
+    logger.debug(`Starting searchAll for: "${options.query}"`);
 
     const data: FileData[] = [];
     let res: Partial<EasynewsSearchResponse> = {
@@ -172,8 +177,8 @@ export class EasynewsAPI {
     const MAX_PAGES = parseInt(process.env.MAX_PAGES || '10'); // Safety limit on number of page requests
     const MAX_RESULTS_PER_PAGE = parseInt(process.env.MAX_RESULTS_PER_PAGE || '250'); // Maximum results per page
 
-    console.info(
-      `[EasynewsAPI] Search limits: max ${TOTAL_MAX_RESULTS} results, max ${MAX_PAGES} pages, ${MAX_RESULTS_PER_PAGE} per page`
+    logger.info(
+      `Search limits: max ${TOTAL_MAX_RESULTS} results, max ${MAX_PAGES} pages, ${MAX_RESULTS_PER_PAGE} per page`
     );
 
     let pageNr = 1;
@@ -188,15 +193,11 @@ export class EasynewsAPI {
 
         // If we've already reached our limit, stop fetching
         if (remainingResults <= 0) {
-          console.debug(
-            `[EasynewsAPI] Reached result limit (${TOTAL_MAX_RESULTS}), stopping pagination`
-          );
+          logger.debug(`Reached result limit (${TOTAL_MAX_RESULTS}), stopping pagination`);
           break;
         }
 
-        console.debug(
-          `[EasynewsAPI] Fetching page ${pageNr} with ${optimalPageSize} results per page`
-        );
+        logger.debug(`Fetching page ${pageNr} with ${optimalPageSize} results per page`);
         const pageResult = await this.search({
           ...options,
           pageNr,
@@ -210,46 +211,46 @@ export class EasynewsAPI {
 
         // No more results
         if (newData.length === 0) {
-          console.debug(`[EasynewsAPI] No more results found, stopping pagination`);
+          logger.debug(`No more results found, stopping pagination`);
           break;
         }
 
         // Duplicate detection - stop if first item of new page matches first item of previously fetched data
         if (data.length > 0 && newData[0]?.['0'] === data[0]?.['0']) {
-          console.debug(`[EasynewsAPI] Duplicate results detected, stopping pagination`);
+          logger.debug(`Duplicate results detected, stopping pagination`);
           break;
         }
 
-        console.debug(`[EasynewsAPI] Adding ${newData.length} results from page ${pageNr}`);
+        logger.debug(`Adding ${newData.length} results from page ${pageNr}`);
         data.push(...newData);
 
         // Stop if we've reached our total limit
         if (data.length >= TOTAL_MAX_RESULTS) {
-          console.debug(
-            `[EasynewsAPI] Reached result limit (${TOTAL_MAX_RESULTS}), trimming and stopping`
-          );
+          logger.debug(`Reached result limit (${TOTAL_MAX_RESULTS}), trimming and stopping`);
           // Trim the array to exactly TOTAL_MAX_RESULTS
           data.length = TOTAL_MAX_RESULTS;
           break;
         }
 
+        logger.debug(
+          `Progress: ${data.length}/${TOTAL_MAX_RESULTS} results (${Math.round(
+            (data.length / TOTAL_MAX_RESULTS) * 100
+          )}%)`
+        );
+
         pageNr++;
       }
 
-      console.debug(`[EasynewsAPI] SearchAll complete, returning ${data.length} total results`);
+      logger.debug(`SearchAll complete, returning ${data.length} total results`);
       return { ...res, data } as EasynewsSearchResponse;
     } catch (error) {
       // If we have partial results, return them
       if (data.length > 0) {
-        console.debug(`[EasynewsAPI] Returning ${data.length} partial results due to error`);
-        console.error(
-          `Partial results returned due to error: ${error instanceof Error ? error.message : 'Unknown error'}`
-        );
+        logger.debug(`Returning ${data.length} partial results due to error`);
+        logger.error(`Search error: ${(error as Error).message}`);
         return { ...res, data } as EasynewsSearchResponse;
       }
-      console.debug(
-        `[EasynewsAPI] No results to return due to error: ${error instanceof Error ? error.message : 'Unknown error'}`
-      );
+      logger.debug(`No results to return due to error: ${(error as Error).message}`);
       throw error;
     }
   }
