+++ packages/cloudflare-worker/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "easynews-plus-plus-addon": "file:../addon",
+    "easynews-plus-plus-shared": "file:../shared",
     "hono": "^4.5.3",
     "hono-stremio": "^0.1.1"
   },
